refactor: inline notification deduplication
